Move beat snap grid implementation details to `ScrollingHitObjectComposer`
