+++ osu.Game/Rulesets/Edit/HitObjectComposer.cs
@@ -90,9 +90,6 @@ namespace osu.Game.Rulesets.Edit
 
         protected DrawableRuleset<TObject> DrawableRuleset { get; private set; }
 
-        [CanBeNull]
-        private BeatSnapGrid beatSnapGrid;
-
         protected HitObjectComposer(Ruleset ruleset)
             : base(ruleset)
         {
@@ -209,8 +206,6 @@ namespace osu.Game.Rulesets.Edit
                         }
                     }
                 },
-                // Must be constructed after drawable ruleset above.
-                (beatSnapGrid = CreateBeatSnapGrid()) ?? Empty(),
             };
 
             toolboxCollection.Items = CompositionTools
@@ -275,37 +270,6 @@ namespace osu.Game.Rulesets.Edit
             }
         }
 
-        protected override void UpdateAfterChildren()
-        {
-            base.UpdateAfterChildren();
-
-            updateBeatSnapGrid();
-        }
-
-        private void updateBeatSnapGrid()
-        {
-            if (beatSnapGrid == null)
-                return;
-
-            if (BlueprintContainer.CurrentTool is SelectTool)
-            {
-                if (EditorBeatmap.SelectedHitObjects.Any())
-                {
-                    beatSnapGrid.SelectionTimeRange = (EditorBeatmap.SelectedHitObjects.Min(h => h.StartTime), EditorBeatmap.SelectedHitObjects.Max(h => h.GetEndTime()));
-                }
-                else
-                    beatSnapGrid.SelectionTimeRange = null;
-            }
-            else
-            {
-                var result = FindSnappedPositionAndTime(InputManager.CurrentState.Mouse.Position);
-                if (result.Time is double time)
-                    beatSnapGrid.SelectionTimeRange = (time, time);
-                else
-                    beatSnapGrid.SelectionTimeRange = null;
-            }
-        }
-
         public override Playfield Playfield => drawableRulesetWrapper.Playfield;
 
         public override IEnumerable<DrawableHitObject> HitObjects => drawableRulesetWrapper.Playfield.AllHitObjects;
@@ -335,12 +299,6 @@ namespace osu.Game.Rulesets.Edit
         /// Construct a relevant blueprint container. This will manage hitobject selection/placement input handling and display logic.
         /// </summary>
         protected virtual ComposeBlueprintContainer CreateBlueprintContainer() => new ComposeBlueprintContainer(this);
-
-        /// <summary>
-        /// Construct an optional beat snap grid.
-        /// </summary>
-        [CanBeNull]
-        protected virtual BeatSnapGrid CreateBeatSnapGrid() => null;
 
         /// <summary>
         /// Construct a drawable ruleset for the provided ruleset.
