+++ osu.Game/Rulesets/Edit/ScrollingHitObjectComposer.cs
@@ -1,6 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
@@ -8,9 +9,11 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using osu.Game.Configuration;
 using osu.Game.Graphics.UserInterface;
+using osu.Game.Rulesets.Edit.Tools;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.UI.Scrolling;
 using osu.Game.Screens.Edit.Components.TernaryButtons;
+using osu.Game.Screens.Edit.Compose.Components;
 using osuTK;
 
 namespace osu.Game.Rulesets.Edit
@@ -20,6 +23,13 @@ namespace osu.Game.Rulesets.Edit
     {
         private readonly Bindable<TernaryState> showSpeedChanges = new Bindable<TernaryState>();
         private Bindable<bool> configShowSpeedChanges = null!;
+
+        private BeatSnapGrid? beatSnapGrid;
+
+        /// <summary>
+        /// Construct an optional beat snap grid.
+        /// </summary>
+        protected virtual BeatSnapGrid? CreateBeatSnapGrid() => null;
 
         protected ScrollingHitObjectComposer(Ruleset ruleset)
             : base(ruleset)
@@ -56,6 +66,42 @@ namespace osu.Game.Rulesets.Edit
                     toggleRuleset.ShowSpeedChanges.Value = enabled;
                     configShowSpeedChanges.Value = enabled;
                 }, true);
+            }
+
+            beatSnapGrid = CreateBeatSnapGrid();
+
+            if (beatSnapGrid != null)
+                AddInternal(beatSnapGrid);
+        }
+
+        protected override void UpdateAfterChildren()
+        {
+            base.UpdateAfterChildren();
+
+            updateBeatSnapGrid();
+        }
+
+        private void updateBeatSnapGrid()
+        {
+            if (beatSnapGrid == null)
+                return;
+
+            if (BlueprintContainer.CurrentTool is SelectTool)
+            {
+                if (EditorBeatmap.SelectedHitObjects.Any())
+                {
+                    beatSnapGrid.SelectionTimeRange = (EditorBeatmap.SelectedHitObjects.Min(h => h.StartTime), EditorBeatmap.SelectedHitObjects.Max(h => h.GetEndTime()));
+                }
+                else
+                    beatSnapGrid.SelectionTimeRange = null;
+            }
+            else
+            {
+                var result = FindSnappedPositionAndTime(InputManager.CurrentState.Mouse.Position);
+                if (result.Time is double time)
+                    beatSnapGrid.SelectionTimeRange = (time, time);
+                else
+                    beatSnapGrid.SelectionTimeRange = null;
             }
         }
     }
